requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.1
schedule==1.2.0
python-dotenv==1.0.0
tenacity==8.2.3
python-dateutil==2.8.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
bcrypt==4.1.1
pyjwt==2.8.0
stripe==7.8.0
python-multipart==0.0.6
